camera: make initial values match default breakpoint
The default breakpoint SingleVertical is not explicitly
set on start-up, instead the UI relies on the initial
values being the ones that would be set for it.

If the breakpoint on start-up is determined to be this
default breakpoint – SingleVertical – without this fix,
the window is left in a state with no start window controls.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>