aperture: Use our preference for caps
Prefere the caps we already compute as being best. This notably makes us
prefer jpeg encoded streams on many cameras, allowing higher frame
rates.

On top of that, ensure we still signal support other caps. This is
important if the camera is already running because of another app -
e.g. Cheese. Right now that app has authority over the chosen mode and
if that doesn't match one of our preferred ones we'd break.