aperture: Handle rotations and conversions in individual pipelines
So we can use native rotation support in gtk4paintablesink and meta-data
support for images and videos. Unfortunately this does not work yet for
the later - to be fixed - however most affected devices are too slow to
handle the rotation during recording anyway.

We can do so by setting individual "filters" for the different pipelines.